requirements.txt
streamlit
groq
python-dotenv
chromadb
bcrypt
